Move component HospitalizedNumberCard from 'components/cards' to 'components/index/CardsMonitoring'
