+++ components/index/CardsMonitoring/HospitalizedNumberCard.vue
